feat: new 'position' option 'Replace [toc] Placeholder'
For backwards compatibility with content that uses the `[toc]` shortcode
from the old "Table of Content Plus" plugin, an new 'position' option is
introduced.

If that option is selected on the plugin admin page, the TOC is inserted
at the place where the old `[toc]` shortcode is found. This also works
for shortcodes with optional arguments, e. g. `[toc exclude="foo*"]`.